pyproject.toml
[project]
name = "LatexOCR"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "DawnMagnet", email = "dawnmagnet@qq.com" }]
dependencies = [
    "pix2tex>=0.1.2",
    "pillow>=10.2.0",
    "fastapi>=0.109.2",
    "uvicorn>=0.27.1",
    "python-multipart>=0.0.9",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
